Drop the in-process Isaac Sim import from the teleop install test
The import exited by signal in the CI container with no traceback and no
stderr, so it reported nothing actionable. Isaac Sim usability from the
uv path is already covered by test_uv_run_isaacsim_trains_cartpole, and
that the extras satisfy every isaaclab_* import the teleop scripts make
is asserted without an install by cli/test_teleop_entrypoints.py.

What remains verifies the teleop extra resolves and the isaaclab teleop
entry point runs from it. Rename the file and class to match, since
neither described the remaining test any more.
